clothing_recommendation_app/requirements.txt
fastapi
uvicorn
jinja2
pandas
numpy
scikit-learn
joblib
pillow
python-multipart
